* BCreated a bow handler
